+++ ML-Frameworks/pytorch-aarch64/get-source.sh
@@ -19,10 +19,10 @@
 
 set -eux -o pipefail
 
-PYTORCH_HASH=fda43c98d1bc82eeb493f6e79a1fd5f1636474a5 # From viable/strict
-IDEEP_HASH=77d4b35c685eecfc8f32ced5381052483bdc3b1d   # From ideep_pytorch
-ONEDNN_HASH=316287183e2503a9db2d975dd04359728aabceb2 # From main
-ACL_HASH=6acccf1730b48c9a22155998fc4b2e0752472148     # From main
+PYTORCH_HASH=8d4926e30a944320adf434016129cb6788eff79b  # From viable/strict
+IDEEP_HASH=e026f3b0318087fe19e2b062e8edf55bfe7a522c    # From ideep_pytorch
+ONEDNN_HASH=0fd3b73a25d11106e141ddefd19fcacc74f8bbfe   # From main
+ACL_HASH=6acccf1730b48c9a22155998fc4b2e0752472148      # From main
 
 function git-shallow-clone {
     (
@@ -53,7 +53,7 @@ function apply-github-patch {
     # Look in the full repo instead.
     # If it can't be found, this time curl will error
     if [[ ! -s $3.patch ]]; then
-       >&2 echo "Commit $2 not found in PR $1. Checking the full repository..."
+       >&2 echo "Commit $3 not found in $1/pull/$2. Checking the full repository..."
        curl --silent --fail -L $1/commit/$3.patch -o $3.patch
     fi
 
@@ -79,12 +79,18 @@ git-shallow-clone https://github.com/pytorch/pytorch.git $PYTORCH_HASH
     # Bump OpenBLAS version. Note that install_openblas.sh has to be rerun in the PyTorch builder Docker container
     sed -i -e 's/v0.3.25/v0.3.28/g' .ci/docker/common/install_openblas.sh
 
+    apply-github-patch https://github.com/pytorch/pytorch 143190 bdb74e24bfc70240fd2260dd7613246d7972fac1 # Enable AArch64 CI scripts to be used for local dev
+
     apply-github-patch https://github.com/pytorch/pytorch 139887 eff3c11b1a31f725b50020ce32f6eddba17b5a94 # Use s8s8s8 for qlinear on aarch64 instead of u8s8u8 with mkl-dnn
     apply-github-patch https://github.com/pytorch/pytorch 139753 16d397416abc44005fc66e377d4d15a0d6131a32 # Add SVE implementation for 8 bit quantized embedding bag on aarch64
     apply-github-patch https://github.com/pytorch/pytorch 136850 6d5aaff8434203f870d76d840158d6989ddd61d0 # Enable XNNPACK for quantized add
     apply-github-patch https://github.com/pytorch/pytorch 140233 6d0b4448bfe3771e076e5c7758333f98810605c4 # Enables static quantization for aarch64
     apply-github-patch https://github.com/pytorch/pytorch 135058 511af4efb5c008a75a196c525a7ad546a9915fd0 # Pass ideep:lowp_kind to matmul_forward::compute on cache misses
-    apply-github-patch https://github.com/pytorch/pytorch 143190 bdb74e24bfc70240fd2260dd7613246d7972fac1 # Enable AArch64 CI scripts to be used for local dev
+    apply-github-patch https://github.com/pytorch/pytorch 142391 8373846f441381a56e7abd905af84102aa52fc7b # parallelize sort
+    apply-github-patch https://github.com/pytorch/pytorch 139387 e5e5d29d6bab882540e36e44a3a75bd187fcbb62 # Add prepacking for linear weights
+    apply-github-patch https://github.com/pytorch/pytorch 139387 19423aaa1af154e1d47d8acf1e677dff727da5aa # Add prepacking for linear weights
+    apply-github-patch https://github.com/pytorch/pytorch 140159 9463c3261f57c42a952f1ba95633833cb1c561fc # cpu: aarch64: enable gemm-bf16f32
+
     git submodule sync
     git submodule update --init --checkout --force --recursive --jobs=$(nproc)
     (
@@ -96,7 +102,7 @@ git-shallow-clone https://github.com/pytorch/pytorch.git $PYTORCH_HASH
             cd mkl-dnn
             git fetch origin $ONEDNN_HASH && git clean -f && git checkout -f FETCH_HEAD
             apply-github-patch https://github.com/oneapi-src/oneDNN 2194 c22f4ae50002ef0a93bfe1895684f36abd92517d # src: cpu: aarch64: lowp_matmul: Make weights constant
-            # Two commits from one PR
+            # Multiple commits from one PR
             apply-github-patch https://github.com/oneapi-src/oneDNN 2212 6a77e84feb442964c91a0101d58fe1473566b185 # src: cpu: aarch64: Enable matmul static quantisation.
             apply-github-patch https://github.com/oneapi-src/oneDNN 2212 efad4f6582c13823d81c78130ab80db57b1381eb # src: cpu: aarch64: Enable convolution static quantisation.
             apply-github-patch https://github.com/oneapi-src/oneDNN 2212 0358abf98dd6c5221a0c40ea47f0a23a1e6cf28e # src: cpu: aarch64: lowp_matmul: Make weights constant